requirements.txt
# Core RAG
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
ollama>=0.2.0

# Streamlit UI
streamlit>=1.35.0
audio-recorder-streamlit>=0.0.8

# Voice
faster-whisper>=1.0.0
pyttsx3>=2.90

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0
